Delete redundant files
