--- src/main/resources/language/en.yml
@@ -1,12 +1,0 @@
-help: "&7- &a/hardcore help &f- &7Get Help\n&7- &a/hardcore on &f- &7Turn on the hardcore mode\n&7- &a/hardcore off &f- &7Turn off the hardcore mode\n&7- &a/hardcore reset &f- &7Reset players' survival states"
-enable: "Hardcore mode is on."
-disable: "Hardcore mode is off."
-reset: "Player state has reset."
-welcome: "WELCOME, NEW PLAYER!"
-welcomeSub: "You only have one life and do your best to survive!"
-dead: "YOU HAVE DIED IN THIS SEASON!"
-deadSub: "Please wait for the reset!"
-alive: "YOU ARE ALIVE!"
-aliveSub: "Survive and earn more points!"
-noperm: "No permission."
-emptyparams: "Empty parameters."
