--- src/main/resources/language/zh.yml
@@ -1,12 +1,0 @@
-help: "&7- &a/hardcore help &f- &7获得帮助\n&7- &a/hardcore on &f- &7打开极限模式\n&7- &a/hardcore off &f- &7关闭极限模式\n&7- &a/hardcore reset &f- &7重置玩家的生存状态"
-enable: "极限模式已开启."
-disable: "极限模式已关闭."
-reset: "玩家状态重置."
-welcome: "欢迎, 新玩家!"
-welcomeSub: "你只有一次生命，尽你所能去生存吧!"
-dead: "你死了!"
-deadSub: "请等待重置!"
-alive: "你还活着!"
-aliveSub: "生存并获得更多积分!"
-noperm: "没有权限."
-emptyparams: "缺少参数."
